packages/docs-test/requirements.txt
supermemory
openai
anthropic
httpx
python-dotenv
